+++ Normalizer/IntegerNormalizer.php
@@ -1,0 +1,52 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the SwaggerResolverBundle package.
+ *
+ * (c) Viktor Linkin <adrenalinkin@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Linkin\Bundle\SwaggerResolverBundle\Normalizer;
+
+use Closure;
+use EXSyst\Component\Swagger\Schema;
+use Linkin\Bundle\SwaggerResolverBundle\Exception\NormalizationFailedException;
+use Symfony\Component\OptionsResolver\Options;
+use function is_numeric;
+
+/**
+ * @author Viktor Linkin <adrenalinkin@gmail.com>
+ */
+class IntegerNormalizer implements SwaggerNormalizerInterface
+{
+    /**
+     * {@inheritdoc}
+     */
+    public function supports(Schema $propertySchema, string $propertyName, bool $isRequired, array $context = []): bool
+    {
+        return $propertySchema->getType() === 'integer';
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getNormalizer(Schema $propertySchema, string $propertyName, bool $isRequired): Closure
+    {
+        return function (Options $options, $value) use ($isRequired, $propertyName) {
+            if (is_numeric($value)) {
+                return (int) $value;
+            }
+
+            if (!$isRequired && $value === null) {
+                return null;
+            }
+
+            throw new NormalizationFailedException($propertyName, (string) $value);
+        };
+    }
+}